# Optimize asset list endpoint